move store and deliver under respective workflow
