--- cg/cli/deliver/__init__.py
@@ -1,1 +1,0 @@
-"""cg cli deliver command and sub commands"""
