--- cg/cli/store/__init__.py
@@ -1,1 +1,0 @@
-"""cg cli store command and sub commands"""
