+++ cg/cli/workflow/mip_rna/mip_rna.py
@@ -8,6 +8,7 @@ from cg.apps import hk, tb, lims
 from cg.apps.environ import environ_email
 from cg.apps.mip import MipAPI
 from cg.apps.mip.fastq import MipFastqHandler
+from cg.cli.workflow.mip_rna import store, deliver
 from cg.cli.workflow.workflow import get_links
 from cg.meta.workflow.mip_rna import AnalysisAPI
 from cg.meta.deliver.mip_rna import MipRnaDeliverAPI as DeliverAPI
@@ -111,3 +112,7 @@ def config_case(context: click.Context, case_id: str, dry: bool = False):
         # Write to trailblazer root dir / case_id
         out_path = context.obj['tb'].save_config(config_data)
         LOG.info("saved config to: %s", out_path)
+
+
+mip_rna.add_command(store)
+mip_rna.add_command(deliver)
